backend/requirements.txt
# FastAPI - Web framework for building the API
fastapi==0.115.0

# Uvicorn - Server that runs the FastAPI app
uvicorn[standard]==0.30.6

# Pydantic - Validates incoming data (makes sure requests have correct format)
pydantic==2.9.2

# Python-multipart - Handles file uploads (for PDF/Word resumes)
python-multipart==0.0.9

# Httpx - Makes HTTP requests (used to call Daniyal's AI service)
httpx==0.27.0

# Python-dotenv - Loads settings from .env file
python-dotenv==1.0.1

# PyPDF2 - Extracts text from PDF files
PyPDF2==3.0.1

# python-docx - Extracts text from Word (.docx) files
python-docx==1.1.2
